seeding for sampling with a np.random.RandomState
Setting `np.seed()` outside of this implementation is not convenient
and can provide a limitation. This commit renders that unnecessary by
maintaining a `np.random.RandomState` object whereever we generate
random values. The user interface is comfortable, since he only needs
to provide an integer (or None) seed and does not need to deal with
streams himself.